added a survey code attribute to json and created a wrapper and heper class to connect the json created to the controller
I did this so I would be able to store the answers into the correct survey repository in the controller

+++ src/main/java/com/surveyMonkey/util/StoreAnswerHelper.java
@@ -3,14 +3,14 @@ package com.surveyMonkey.util;
 import java.util.List;
 
 public class StoreAnswerHelper {
-    private int questionId;
+    private long questionId;
     private String answer;
 
     public void setQuestionId(int questionId){
         this.questionId = questionId;
     }
 
-    public int getQuestionId(){
+    public long getQuestionId(){
         return questionId;
     }
